llms-os-project/llms-os/requirements.txt
# Core dependencies
pyyaml==6.0.1
jinja2==3.1.2
requests==2.31.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# API clients
openai==0.28.0
anthropic==0.8.1

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
tenacity==8.2.3

# Validation
pydantic==2.5.3
jsonschema==4.20.0
